requirements.txt.txt
flask
python-pptx
python-docx
langchain
faiss-cpu
sentence-transformers
huggingface-hub
